refactor(components): migrate CodeEditor from TSX to Vue SFC
Convert CodeEditor and MonacoEditor components from TypeScript/JSX to Vue Single File Components for better Vue ecosystem integration and maintainability. Update package dependencies to use workspace catalog for consistent versioning across packages.

- Replace defineComponent with <script setup> syntax in CodeEditor.vue and MonacoEditor.vue
- Remove index.tsx and MonacoEditor.tsx files as they're no longer needed
- Add proper barrel export in new index.ts file
- Update web-app and auth-app to import CodeEditor from components package directly
- Align Vue and related dependency versions across workspace using pnpm catalog
