+++ frontend/packages/auth-app/package.json
@@ -11,16 +11,16 @@
   "dependencies": {
     "@rpa/components": "workspace:*",
     "@rpa/shared": "workspace:*",
-    "vue": "^3.5.24"
+    "vue": "catalog:"
   },  
   "devDependencies": {
     "@types/node": "^24.10.0",
-    "@vitejs/plugin-vue": "^6.0.1",
-    "@vitejs/plugin-vue-jsx": "^5.0.1",
+    "@vitejs/plugin-vue": "catalog:",
+    "@vitejs/plugin-vue-jsx": "catalog:",
     "@vue/tsconfig": "^0.8.1",
     "tailwindcss": "^3.4.4",
     "typescript": "~5.9.3",
-    "vite": "npm:rolldown-vite@7.2.2",
+    "vite": "catalog:",
     "vue-tsc": "^3.1.3"
   }
 }
