+++ frontend/packages/components/src/components/CodeEditor/index.ts
@@ -1,0 +1,3 @@
+import { defineAsyncComponent } from 'vue'
+
+export const CodeEditor = defineAsyncComponent(() => import('./CodeEditor.vue'))
