+++ frontend/packages/web-app/src/views/Arrange/components/flow/FlowCode.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-import { useTheme } from '@rpa/components'
+import { useTheme, CodeEditor } from '@rpa/components'
 import { storeToRefs } from 'pinia'
-import { defineAsyncComponent, onBeforeMount, onUnmounted } from 'vue'
+import { onBeforeMount, onUnmounted } from 'vue'
 
 import { getRootBaseURL } from '@/api/http/env'
 import { useProcessStore } from '@/stores/useProcessStore'
@@ -12,7 +12,6 @@ const { isDark } = useTheme()
 const { pyCodeText } = storeToRefs(processStore)
 
 const baseUrl = `${getRootBaseURL()}/scheduler`
-const CodeEditor = defineAsyncComponent(() => import('@rpa/components').then(m => m.CodeEditor))
 
 function handleUpdate(codeString: string) {
   processStore.setCodeText(codeString)
