+++ frontend/pnpm-workspace.yaml
@@ -8,10 +8,11 @@ catalog:
   '@vitejs/plugin-vue': ^6.0.3
   '@vitejs/plugin-vue-jsx': ^5.1.3
   ant-design-vue: ^4.2.6
-  vite: 8.0.0-beta.5
+  vue: "^3.5.24"
+  vite: ^8.0.0-beta.10
   tailwind-merge: ^3.4.0
   tailwindcss: "^3.4.4"
-  tsdown: "^0.19.0-beta.2"
+  tsdown: "^0.20.1"
 
 ignoredBuiltDependencies:
   - '@bundled-es-modules/glob'
